Refactor: move logic to DependencyHolder
